Tower shooting

+++ Assets/Content/Scene Game/Scripts/IDamageable.cs
@@ -3,5 +3,5 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface IDamageable {
-	void Damage(float amount);
+	void Damage(int amount);
 }
